requirements.txt
numpy==1.23.5
pandas==1.5.3
Pillow==9.4.0
streamlit==1.20.0
streamlit_option_menu==0.3.2
tensorflow==2.11.0
tensorflow_intel==2.12.0rc1
